requirements.txt
dash==1.7.0
pandas==0.25.3
plotly-express==0.4.1

# production
gunicorn==20.0.4
